Completed DistanceMap
Not too difficult to implement, only three short methods including the initializer. What this class does is create an empty private dictionary called _distances, we use the method add_distance to add the two cities as a key value to the dictionary and make it refer to the distance between them, but if there is no parameter for distance thats passed, we just add -1 as the distance by default. Method distance just loops over the dictionary until it finds the cities that its looking for and returns the distance value, but it returns None if it doesn't find it.